requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.4
python-dotenv>=1.0
pytesseract>=0.3.10
Pillow>=10.0
pandas>=2.2
requests>=2.32
gunicorn>=21.2
